+++ components/bsa/bsafile.cpp
@@ -246,7 +246,7 @@ void Bsa::BSAFile::writeHeader()
     {
         auto& f = mFiles[i];
         offsets[i * 2] = f.mFileSize;
-        offsets[i * 2 + 1] = f.mOffset - fileDataOffset;
+        offsets[i * 2 + 1] = static_cast<uint32_t>(f.mOffset - fileDataOffset);
         offsets[2 * filenum + i] = f.mNameOffset;
         hashes[i] = f.mHash;
     }
@@ -312,7 +312,7 @@ void Bsa::BSAFile::addFile(const std::string& filename, std::istream& file)
     newFile.mHash = getHash(filename);
 
     if (mFiles.empty())
-        newFile.mOffset = static_cast<uint32_t>(newStartOfDataBuffer);
+        newFile.mOffset = static_cast<uint64_t>(newStartOfDataBuffer);
     else
     {
         std::vector<char> buffer;
@@ -325,7 +325,7 @@ void Bsa::BSAFile::addFile(const std::string& filename, std::istream& file)
             stream.read(buffer.data(), firstFile.mFileSize);
 
             stream.seekp(0, std::ios::end);
-            firstFile.mOffset = static_cast<uint32_t>(stream.tellp());
+            firstFile.mOffset = static_cast<uint64_t>(stream.tellp());
 
             stream.write(buffer.data(), firstFile.mFileSize);
 
@@ -333,7 +333,7 @@ void Bsa::BSAFile::addFile(const std::string& filename, std::istream& file)
             std::rotate(mFiles.begin(), mFiles.begin() + 1, mFiles.end());
         }
         stream.seekp(0, std::ios::end);
-        newFile.mOffset = static_cast<uint32_t>(stream.tellp());
+        newFile.mOffset = static_cast<uint64_t>(stream.tellp());
     }
 
     newFile.mNameOffset = static_cast<uint32_t>(mStringBuf.size());
